refactor: merged into one card component
- 다른 컴포넌트와 일관성을 유지하기 위해서 별개로 분리된 파일을
하나로 합침
- time-function util로 이동되었기 때문에 불필요한 파일 제거
- 조회수로 변경했기 때문에 채팅아이콘에서 조회수 아이콘으로 변경
